Fix: Fixed bug in book contract

+++ src/main/java/com/book/store/Repository/BookRepository.java
@@ -10,11 +10,11 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Integer> {
    Optional<Book>  findByTitle(String title);
 
     @Query("SELECT b.description FROM Book b WHERE b.id = :id")
-   String  getDescriptionById(Long id);
+   String  getDescriptionById(Integer id);
   // Book findById(Long id);
